add user delete

+++ blog/views.py
@@ -8,6 +8,7 @@ from .forms import QuestionForm, AnswerForm, QuestionImageForm, AnswerImageForm,
 from django.shortcuts import get_object_or_404
 from django.db import transaction
 from django.contrib.auth.decorators import login_required
+#from django.contrib import messages
 # Create your views here.
 
 def home(request):
@@ -48,7 +49,7 @@ def detail_question(request, pk):
     answers = Answer.objects.filter(question = pk)
 
     if request.method == "POST":
-        answer_form = AnswerForm(request.POST, request.FILES)
+        answer_form = AnswerForm(request.POST)#, request.FILES)
         image_formset = AnswerImageFormSet(request.POST, request.FILES)
 
         if answer_form.is_valid() and image_formset.is_valid():
@@ -71,5 +72,32 @@ def detail_question(request, pk):
 
 def question_remove(request, pk):
     question=get_object_or_404(Question, pk=pk)
-    question.delete()
-    return redirect('home')
+    
+    if request.user != question.author: #and not request.user.is_staff
+        #messages.warning(request, '권한 없음') 이것저것 설치해야함
+        return redirect('detail_question', pk=pk)
+    else :
+        question.delete()
+        return redirect('home')
+"""
+def question_update(request, pk):
+    question = get_object_or_404(Question, pk=pk)
+
+    if request.user != question.author:
+        #messages.warning(request, '권한 없음') 이것저것 설치해야함
+        return redirect('detail_question')
+    if request.method == "POST":
+        form = QuestionForm(request.POST, instance = question)
+        if form.is_vaild():
+            form.save()
+            return redirect(question)
+    else:
+        form = QuestionForm(instance=question)
+    return render(request, 'detail_question.html', {'form':form})
+
+
+def answer_remove(request, pk):
+    answer=get_object_or_404(Answer, pk=pk)
+    answer.delete()
+    return redirect('detail_question')
+"""
